Simplify edit() by removing the `append` option.
This was always somewhat problematic, given that it takes a zero-width original
span, which could occur at the end point of multiple items.

Given the single-item sequence

```
- a
```

doing an `s.append("b")` and `del s[0]` should not remove the `b` edit, but
doing `s.anneal()` should.  Now that broad compatibility is better, the
`items.py` code does not use `append=True` anymore and we don't need this path.